pyproject.toml
[tool.poetry]
name = "knitpkg-mt"
version = "0.1.0"
description = "KnitPkg for MetaTrader – Professional package manager for MQL5/MQL4"
authors = ["Douglas Rechia <rechia@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "knitpkg"}]

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.12.5"
gitpython = "^3.1.45"
rich = "^14.2.0"
click = "^8.3.1"
typer = {extras = ["all"], version = "^0.20.0"}
packaging = "^25.0"
chardet = "^5.2.0"
pyyaml = "^6.0.3"
jinja2 = "^3.1.6"
httpx = "^0.28.1"
keyring = "^25.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
pytest-xdist = "^3.8.0"
pytest-sugar = "^1.1.1"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
test = "pytest"
tests = "pytest -xvs"
cov = "pytest --cov=knitpkg --cov-report=term-missing --cov-report=html:htmlcov"
ci = "pytest --junitxml=reports/junit.xml --cov=knitpkg --cov-report=xml"
watch = "pytest-watch -- -xvs"
manifest = "pytest tests/test_manifest.py -v"

[tool.poetry.scripts]
kp-mt = "knitpkg.cli:app"
